simplify

+++ misc/sparse_matrix.py
@@ -7,12 +7,12 @@ n = 8
 K = ti.linalg.SparseMatrixBuilder(n,
                                   n,
                                   max_num_triplets=100,
-                                  dtype=ti.f32,
+                                  dtype=ti.f64,
                                   storage_format='col_major')
 f = ti.linalg.SparseMatrixBuilder(n,
                                   1,
                                   max_num_triplets=100,
-                                  dtype=ti.f32,
+                                  dtype=ti.f64,
                                   storage_format='col_major')
